Added new manage.py command and added more functionality to import_movies
There is now a functioanlity for manage.py that lets you clear all the curren tables.

Also a first version of the movie importer is online.

It has everything except the ratings (That we will probably do in milestone 2) .
Adding to database only definitely works, if it is empty before, so if you added data, use the flush function.